app/run-requirements.txt
selenium==4.15.2
pandas==2.1.3
pyyaml==6.0.1
flask==3.0.0
mysql-connector-python==8.2.0